backend/requirements.txt
fastapi
uvicorn
redis[hiredis]
boto3
requests-oauthlib
cryptography
python-dotenv
pydantic
email-validator
pyjwt
passlib
tzdata
pytest
black
isort
flake8
mypy
python-jose
requests
pandas
numpy
python-multipart
jq
typer